requirements.txt
asgiref==3.9.1
bcrypt==4.3.0
CacheControl==0.14.3
cachetools==6.2.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==46.0.2
Django==5.2.5
django-cors-headers==4.7.0
django-jazzmin==3.0.1
django-rest-framework==0.1.0
djangorestframework==3.16.1
firebase-admin==6.4.0
google-api-core==2.25.1
google-api-python-client==2.184.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.4.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.75.1
grpcio-status==1.75.1
httplib2==0.31.0
idna==3.10
msgpack==1.1.1
pillow==11.3.0
proto-plus==1.26.1
protobuf==6.32.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.10.1
PyMySQL==1.1.2
pyOpenSSL==25.3.0
pyparsing==3.2.5
reportlab==4.4.5
requests==2.32.5
rsa==4.9.1
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
qrcode[pil]==7.4.2
openpyxl==3.1.2
pandas>=2.2.3
pyshp==2.3.1
pyproj==3.6.1
channels>=4.0.0
daphne>=4.0.0